refactor(frontend): move Secrets console to Workspace group + /workspace/secrets
The zero-knowledge secret store is a workspace-scoped store (like resources/
storage), not a configuration surface — move its sidebar entry from Settings to
the Workspace group and its route from /workspace/settings/secrets to
/workspace/secrets. Old path keeps a back-compat redirect. Page content is
unchanged (git mv); it already used the workspace i18n namespace.
